refactor: move transaction utils to its own file
